requirements.txt
Pillow>=9.2.0
playwright>=1.25.2
py-cord>=2.1.1